fix: correct hms saturation phase and coriolis output
- use liquid saturation mixing ratio when computing hms

- cast parsed Coriolis parameter to a plain float

- bump pyvvm version to 0.4.2

+++ src/pyvvm/config.py
@@ -106,7 +106,7 @@ def parse_vvm_setup(path: str | Path) -> dict[str, Any]:
     # Calculate Coriolis parameter if CORIOLIS is enabled
     if 'CORIOLIS' in config['defines'] and 'RLAT' in config:
         rlat_rad = np.radians(config['RLAT'])
-        config['f'] = 2 * omega * np.sin(rlat_rad)
+        config['f'] = float(2 * omega * np.sin(rlat_rad))
     else:
         config['f'] = None
